pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Using Codex & Claude Code to create a personal website"
authors = [
    {name = "Ben"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
exclude = ["__pycache__", "*.pyc", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=personal_website --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["personal_website"]
